backend/requirements.txt
aiocache==0.12.3
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.2
aiosqlite==0.20.0
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.3.0
bcrypt==4.0.1
beanie==1.26.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.0
fastapi-cli==0.0.5
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
lazy-model==0.2.0
loguru==0.7.2
Mako==1.3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
motor==3.5.1
multidict==6.1.0
orjson==3.10.12
passlib==1.7.4
pillow==11.0.0
propcache==0.2.1
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic-extra-types==2.10.0
pydantic-settings==2.5.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.10.1
pymongo==4.8.0
python-dotenv==1.0.1
python-multipart==0.0.19
PyYAML==6.0.2
redis==5.2.1
requests==2.32.0
rich==13.9.4
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
toml==0.10.2
transliterate==1.10.2
typer==0.13.1
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.3.0
uvicorn==0.31.0
watchfiles==0.24.0
websockets==14.1
win32-setctime==1.1.0
yarl==1.18.3
